requirements.txt
huggingface-hub=0.32.4
python-dotenv=1.0.1
